Mahesh

+++ portfolio/src/main/java/com/portfolio/dao/UserJpa.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.portfolio.entity.User;
 
-public interface UserJpa extends JpaRepository<User, Long>{
+public interface UserJpa extends JpaRepository<User, Integer>{
 	User findById(int id);
 	User findByUserName(String userName);
